requirements.txt
aiohttp==0.21.5
beautifulsoup4==4.4.1
chardet==2.3.0
Django==1.10.2
django-appconf==1.0.2
django-bootstrap3==7.1.0
django-ckeditor==5.0.3
django-classy-tags==0.7.2
django-formtools==1.0
django-grappelli==2.8.1
django-imagekit==3.3
django-modelcluster==2.0
django-modeltranslation==0.12
django-mptt==0.8.4
django-sekizai==0.9.0
django-taggit==0.18.3
django-tastypie==0.13.3
django-tinymce==2.4.0
django-translatable==0.1.1
django-transmeta==0.7.3
django-treebeard==4.0.1
django-webpack-loader==0.2.4
django-widget-tweaks==1.4.1
djangocms-admin-style==1.2.2
djangorestframework==3.3.3
FeinCMS==1.12.1
html5lib==0.9999999
lxml==3.6.0
mysqlclient==1.3.7
numpy==1.11.3
pilkit==1.1.13
Pillow==3.2.0
pymongo==3.2.2
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2016.4
qrcode==5.3
requests==2.9.1
selenium==3.0.2
six==1.10.0
solver==0.0.4
telepot==6.8
Unidecode==0.4.19
virtualenv==15.0.1
wagtail==1.4.5
Willow==0.3.1
xmlrunner==1.7.7
